Move auto-connect code into autoConnect.ts. Leave multiple connection handling to backend.
