+++ lynxkite-app/web/src/workspace/autoConnect.ts
@@ -1,26 +1,19 @@
 // Auto-connect logic extracted from Workspace.tsx for readability.
 
-import type { Edge, Node, XYPosition } from "@xyflow/react";
+import { type Edge, type Node, useReactFlow, type XYPosition } from "@xyflow/react";
+import { type MouseEvent, useCallback, useMemo, useState } from "react";
+import type { Workspace, WorkspaceEdge } from "../apiTypes.ts";
 import { getHandles } from "./nodes/LynxKiteNode.tsx";
 
-export const MIN_AUTO_CONNECT_DISTANCE = 220;
+const MIN_AUTO_CONNECT_DISTANCE = 220;
 
-export function allowsMultipleConnections(_inputType: unknown): boolean {
-  // Assume all inputs allow multiple connections for now.
-  // If it's annoying, we can switch to assuming nothing allows multiple connections.
-  return true;
-}
-
-export function getNodeHandles(node: Node, edges: Edge[]) {
+function getNodeHandles(node: Node, edges: Edge[]) {
   const inputs = (node.data as any)?.meta?.inputs ?? [];
   const outputs = (node.data as any)?.meta?.outputs ?? [];
-  return getHandles({ edges }, node.id, inputs, outputs).map((h) => ({
-    ...h,
-    acceptsMultipleConnections: allowsMultipleConnections(undefined),
-  }));
+  return getHandles({ edges } as Workspace, node.id, inputs, outputs);
 }
 
-export function getHandlePosition(
+function getHandlePosition(
   absPos: XYPosition,
   node: Node,
   handle: { position: string; offsetPercentage?: number },
@@ -34,7 +27,7 @@ export function getHandlePosition(
   return { x: absPos.x + width * offset, y: absPos.y + height };
 }
 
-export function findClosestHandlePair(
+function findClosestHandlePair(
   sourceNode: Node,
   targetNode: Node,
   edges: Edge[],
@@ -50,12 +43,6 @@ export function findClosestHandlePair(
   for (const sh of sourceHandles) {
     const sp = getHandlePosition(getAbsPos(sourceNode), sourceNode, sh);
     for (const th of targetHandles) {
-      const isOccupied = edges.some(
-        (e) => e.target === targetNode.id && e.targetHandle === th.name,
-      );
-      if (isOccupied && !th.acceptsMultipleConnections) {
-        continue;
-      }
       const tp = getHandlePosition(getAbsPos(targetNode), targetNode, th);
       const distance = Math.hypot(sp.x - tp.x, sp.y - tp.y);
       if (distance < bestDistance) {
@@ -67,7 +54,7 @@ export function findClosestHandlePair(
   return bestPair;
 }
 
-export function edgeExists(
+function edgeExists(
   source: string,
   sourceHandle: string,
   target: string,
@@ -81,4 +68,111 @@ export function edgeExists(
       e.target === target &&
       e.targetHandle === targetHandle,
   );
+}
+
+export function useAutoConnect(edges: Edge[], crdt: any) {
+  const reactFlow = useReactFlow();
+  // A preview of the auto-connect edge while dragging a node.
+  const [previewEdge, setPreviewEdge] = useState<Edge | null>(null);
+  const renderedEdges = useMemo(() => {
+    return previewEdge ? [...edges.filter((e) => e.id !== previewEdge.id), previewEdge] : edges;
+  }, [edges, previewEdge]);
+
+  function getAbsPos(node: Node) {
+    const internal = reactFlow.getInternalNode(node.id);
+    return internal?.internals.positionAbsolute ?? node.position;
+  }
+
+  const getClosestEdge = useCallback(
+    (draggedNode: Node) => {
+      const draggedPos = getAbsPos(draggedNode);
+      if (!draggedPos || draggedNode.type === "node_group") {
+        return null;
+      }
+
+      let bestEdge: WorkspaceEdge | null = null;
+      let bestDistance = Number.POSITIVE_INFINITY;
+      for (const n of reactFlow.getNodes()) {
+        if (n.id === draggedNode.id || n.type === "node_group") {
+          continue;
+        }
+        const pos = getAbsPos(n);
+        const closeNodeIsSource = pos.x < draggedPos.x;
+        const sourceNode = closeNodeIsSource ? n : draggedNode;
+        const targetNode = closeNodeIsSource ? draggedNode : n;
+        const bestPair = findClosestHandlePair(sourceNode, targetNode, edges, getAbsPos);
+        if (!bestPair) continue;
+        if (bestPair.distance >= MIN_AUTO_CONNECT_DISTANCE || bestPair.distance >= bestDistance) {
+          continue;
+        }
+        bestDistance = bestPair.distance;
+        bestEdge = {
+          id: `${sourceNode.id} ${bestPair.sourceHandle} ${targetNode.id} ${bestPair.targetHandle}`,
+          source: sourceNode.id,
+          sourceHandle: bestPair.sourceHandle,
+          target: targetNode.id,
+          targetHandle: bestPair.targetHandle,
+        };
+      }
+
+      return bestEdge;
+    },
+    [reactFlow, edges],
+  );
+
+  const onNodeDrag = useCallback(
+    (_event: MouseEvent | TouchEvent, draggedNode: Node) => {
+      const closeEdge = getClosestEdge(draggedNode);
+      if (
+        !closeEdge ||
+        edgeExists(
+          closeEdge.source,
+          closeEdge.sourceHandle!,
+          closeEdge.target,
+          closeEdge.targetHandle!,
+          edges,
+        )
+      ) {
+        if (previewEdge) setPreviewEdge(null);
+        return;
+      }
+      const previewId = `preview:${closeEdge.id}`;
+      if (previewEdge?.id === previewId) return;
+      setPreviewEdge({
+        ...closeEdge,
+        id: previewId,
+        className: "temp-preview-edge",
+        style: {
+          strokeDasharray: "8 6",
+          strokeLinecap: "round",
+        },
+        selectable: false,
+        deletable: false,
+        focusable: false,
+      });
+    },
+    [getClosestEdge, edges, previewEdge],
+  );
+
+  const onNodeDragStop = useCallback(
+    (_event: MouseEvent | TouchEvent, draggedNode: Node) => {
+      const closeEdge = getClosestEdge(draggedNode);
+      setPreviewEdge(null);
+      if (!closeEdge) return;
+      if (
+        edgeExists(
+          closeEdge.source,
+          closeEdge.sourceHandle!,
+          closeEdge.target,
+          closeEdge.targetHandle!,
+          edges,
+        )
+      ) {
+        return;
+      }
+      crdt?.addEdge(closeEdge);
+    },
+    [getClosestEdge, crdt, edges],
+  );
+  return { onNodeDrag, onNodeDragStop, renderedEdges };
 }
